+++ openeo/rest/datacube.py
@@ -1943,6 +1943,7 @@ class DataCube(_ProcessGraphAbstraction):
         :return: None if the result is stored to disk, or a bytes object returned by the backend.
         """
         if format is None and outputfile is not None:
+            # TODO #401/#449 don't guess/override format if there is already a save_result with format?
             format = guess_format(outputfile)
         cube = self._ensure_save_result(format=format, options=options)
         return self._connection.download(cube.flat_graph(), outputfile)
@@ -2062,6 +2063,7 @@ class DataCube(_ProcessGraphAbstraction):
         if "format" in format_options and not out_format:
             out_format = format_options["format"]  # align with 'download' call arg name
         if not out_format and outputfile:
+            # TODO #401/#449 don't guess/override format if there is already a save_result with format?
             out_format = guess_format(outputfile)
 
         job = self.create_job(